fix: fix for nested source paths and sources outside of parent directory
Previously, each level of source file pushed the file name to the path. This caused an error when the source file name was a path, not just the file name. Additonally, this caused an error when the source file was higher up the directory tree (towards the root) than the stowfile. To fix these issues, this commit adds a struct 'SmartPopPathBuf' for pushing/popping multiple directories(+file name) to a path at one time.

Closes issue #1